+++ src/Server/panel.py
@@ -10,6 +10,7 @@ parent_dir = os.path.abspath(os.path.join(current_dir, os.pardir))
 sys.path.append(parent_dir)
 
 from Config import check
+from MainMenu import themes
 
 class ServerPanel(ctk.CTk):
     def __init__(self, server_instance=None):
@@ -40,20 +41,7 @@ class ServerPanel(ctk.CTk):
         self.table.pack(expand=False, fill="both", padx=20, pady=20)
 
     def title_color(self):
-        tb_color = 0x333333
-        if int(check.get_config_value("theme")) == 0:
-            tb_color = 0x333333
-        elif int(check.get_config_value("theme")) == 1:
-            tb_color = 0xFFFFFF
-        else:
-            tb_color = 0x333333
-
-        HWND = windll.user32.GetParent(self.winfo_id())
-        windll.dwmapi.DwmSetWindowAttribute(
-            HWND,
-            35,
-            byref(c_int(tb_color)),
-            sizeof(c_int))
+        themes.title_bar_color_handle(self)
 
     def window(self, fg_cl):
         self.configure(fg_color=fg_cl)
@@ -67,12 +55,12 @@ class ServerPanel(ctk.CTk):
         text_bg = "#4a4a4a"
         text = "white"
         hover_color="#4d4d4d"
-        if int(check.get_config_value("theme")) == 0:
+        if check.get_config_value("theme") == "dark":
             fg_cl = "#2b2b2b"
             text_bg = "#4a4a4a"
             text = "white"
             hover_color="#4d4d4d"
-        elif int(check.get_config_value("theme")) == 1:
+        elif check.get_config_value("theme") == "light":
             fg_cl = "white"
             text_bg = "#f0f0f0"
             text = "black"